# Summary

I have implemented `:git_turbo` strategy for super-fast deployment.

This strategy utilizes Git Worktree, Hardlink, and Rsync.
Following are the procedures in this strategy.

1. Locally, checkout worktree to `.local_repo/git_turbo_worktree`.
2. Locally, copy hardlink files by `cp -al` from `.local_repo/git_turbo_worktree` to `local_release_path`.
3. Remotely, copy hardlink files by `cp -al` from last release to `release_path`.
4. Rsync files from `local_release_path` to remote `release_path`.

The hardlinked files minimize disk I/O usually required for making release directory.
Preserved content and file attributes minimize data transfer by `rsync`.

This strategy is most suitable for continuous deployment where the diff between consecutive releases is small.

# Performance

I have tested the time to deploy a Rails repository to a remote server.

* Approx. File Count: ~ 25,000
* Approx. Total File Size: ~ 2.7G

The result of deploying the same code is as follows:

* `bundle_rsync:create_release`
  * before: 6 sec
  * after: 8 sec
* `bundle_rsync:rsync_release`
  * before: 76 sec
  * after: 2 sec

Since the total cost of `bundle_rsync:rsync_release` is proportional to the number of servers,
30x faster `bundle_rsync:rsync_release` is effective for large-scale deployment.